refactored ts code
