+++ asset-src/ts/change_form_template.ts
@@ -1,3 +1,5 @@
-import { init_live_editor_render } from "./djangocms_equation";
+import { init_live_editor_render } from './function_definitions/change_form'
 
-init_live_editor_render(true);
+document.addEventListener('DOMContentLoaded', function(event) {
+  init_live_editor_render(true)
+})
