+++ asset-src/ts/render_equation.ts
@@ -1,0 +1,5 @@
+import { render_full_page } from './function_definitions/djangocms_equation'
+
+document.addEventListener('DOMContentLoaded', () => {
+  render_full_page()
+})
